Refactor Aspect to separate module
