+++ build.gradle
@@ -14,6 +14,9 @@ configure(subprojects.findAll { it.name != 'sample' }) {
 
   repositories {
     jcenter()
+    maven {
+      url 'https://oss.sonatype.org/content/repositories/snapshots/'
+    }
   }
 
   group = GROUP
