+++ gradle.properties
@@ -1,6 +1,7 @@
 VERSION_NAME=0.1.5-SNAPSHOT
 GROUP=com.github.xfumihiro.view-inspector
 
+ASPECT_ARTIFACT_ID=view-inspector-aspect
 RUNTIME_ARTIFACT_ID=view-inspector-runtime
 PLUGIN_ARTIFACT_ID=view-inspector-plugin
