+++ settings.gradle
@@ -1,1 +1,4 @@
-include ':sample', ':view-inspector-plugin', ':view-inspector-runtime'
+include ':sample'
+include ':view-inspector-plugin'
+include ':view-inspector-runtime'
+include ':view-inspector-aspect'
