+++ view-inspector-aspect/src/main/AndroidManifest.xml
@@ -1,0 +1,7 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="view_inspector.aspect"
+    >
+
+  <uses-sdk android:minSdkVersion="14"/>
+
+</manifest>
