+++ view-inspector-runtime/build.gradle
@@ -1,7 +1,3 @@
-import org.aspectj.bridge.IMessage
-import org.aspectj.bridge.MessageHandler
-import org.aspectj.tools.ajc.Main
-
 apply plugin: 'com.android.library'
 apply plugin: 'com.neenbedankt.android-apt'
 
@@ -15,43 +11,8 @@ android {
   }
 }
 
-android.libraryVariants.all { variant ->
-  JavaCompile javaCompile = variant.javaCompile
-  javaCompile.doLast {
-    String[] args = ["-showWeaveInfo",
-                     "-1.5",
-                     "-inpath", javaCompile.destinationDir.toString(),
-                     "-aspectpath", javaCompile.classpath.asPath,
-                     "-d", javaCompile.destinationDir.toString(),
-                     "-classpath", javaCompile.classpath.asPath,
-                     "-bootclasspath", android.bootClasspath.join(File.pathSeparator)]
-
-    MessageHandler handler = new MessageHandler(true);
-    new Main().run(args, handler)
-
-    def log = project.logger
-    for (IMessage message : handler.getMessages(null, true)) {
-      switch (message.getKind()) {
-        case IMessage.ABORT:
-        case IMessage.ERROR:
-        case IMessage.FAIL:
-          log.error message.message, message.thrown
-          break;
-        case IMessage.WARNING:
-        case IMessage.INFO:
-          log.info message.message, message.thrown
-          break;
-        case IMessage.DEBUG:
-          log.debug message.message, message.thrown
-          break;
-      }
-    }
-  }
-}
-
 dependencies {
   compile 'com.google.dexmaker:dexmaker:1.1'
-  compile 'org.aspectj:aspectjrt:1.8.6'
 
   apt 'com.google.dagger:dagger-compiler:2.0.1'
   compile 'com.google.dagger:dagger:2.0.1'
